Rename README.md to Statics Task
A statistical overview of all columns can be obtained through the.describe() function of the pandas library. This function provides us with the number of valid entries that were entered (count), the number of unique entries that were entered (unique), the entry that was entered the most frequently (top), and the frequency of the entry that was entered the most frequently (freq). Even though the values represent percentages, prior to the data transformation, all of the columns are considered as though they contained categorical data